Move class to jackson package
